Changed Player health and enemy health from int to double. Created battles logic.

+++ RogueMetalicana/RogueMetalicana/Constants/Player/PlayerConstants.cs
@@ -13,7 +13,7 @@
         /// <summary>
         /// Starting stats for the player
         /// </summary>
-        public const int StartingHealth = 100;
+        public const double StartingHealth = 100;
         public const int StartingArmor = 20;
         public const int StartingDamage = 25;
